Change annual income page to use currency

+++ app/pages/AnnualIncomePage.scala
@@ -18,7 +18,7 @@ package pages
 
 import play.api.libs.json.JsPath
 
-case object AnnualIncomePage extends QuestionPage[Int] {
+case object AnnualIncomePage extends QuestionPage[BigDecimal] {
 
   override def path: JsPath = JsPath \ toString
